requirements.txt
msal
python-dotenv
baserowapi
requests
jinja2